Add date helper for tests and rename helpers
